Fix migration

+++ database/migrations/2018_07_17_163723_create_role_users_table.php
@@ -15,7 +15,7 @@ class CreateRoleUsersTable extends Migration
     {
         Schema::create('role_users', function (Blueprint $table) {
             $table->unsignedBigInteger('user_id');
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('role_id');
 
             $table->unique(['user_id','role_id']);
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
